ThanhTung - commit - favorite

+++ zing/src/main/java/com/m2m/zing/api/FavoriteAPI.java
@@ -1,6 +1,7 @@
 
 package com.m2m.zing.api;
 
+import com.google.gson.stream.JsonToken;
 import com.m2m.zing.constant.ModelAttributes;
 import com.m2m.zing.model.Favorite;
 import com.m2m.zing.model.Song;
@@ -12,6 +13,7 @@ import com.m2m.zing.service.UserService;
 import jakarta.servlet.http.HttpSession;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
 import java.time.LocalDate;
@@ -86,12 +88,22 @@ public class FavoriteAPI {
         return ResponseEntity.ok(result);
     }
 
-    @DeleteMapping
-    public ResponseEntity<Map<String, Object>> deleteFavorite(@RequestBody Favorite favorite) {
+    @DeleteMapping("/{songId}")
+    public ResponseEntity<Map<String, Object>> deleteFavorite(@PathVariable Long songId) {
         Map<String, Object> result = new HashMap<>();
         try {
-            favoriteService.deleteFavorite(favorite);
-            result.put("status", "Success");
+            User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+            FavoriteId favoriteId = new FavoriteId();
+            favoriteId.setSong(songId);
+            favoriteId.setUser(user.getUserId());
+            Favorite favorite = favoriteService.getByFavoriteId(favoriteId);
+            if (favorite != null) {
+                favoriteService.deleteFavorite(favorite);
+                result.put("status", "Success");
+            }else{
+                result.put("status", "Failed");
+                result.put("detail","not found favorites with this song and user");
+            }
         } catch (Exception e) {
             result.put("status", "Error");
             result.put("detail", e.toString());
@@ -104,10 +116,7 @@ public class FavoriteAPI {
         Map<String, Object> result = new HashMap<>();
         try {
             Song song = songService.getSongById(songId);
-            User user = (User) httpSession.getAttribute(ModelAttributes.CURRENT_USER);
-            if (user == null) {
-                user = userService.getUserById((long) 1);
-            }
+            User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
             Favorite favorite = new Favorite();
             favorite.setSong(song);
             favorite.setUser(user);
